With all the headerfile changes,  some functions didn't get there arg types fixed up.

+++ libc/misc/sysvipc/sem.c
@@ -72,7 +72,7 @@ int
 semop (semid, sops, nsops)
     int semid;
     struct sembuf *sops;
-    unsigned int nsops;
+    size_t nsops;
 {
     return __ipc(IPCOP_semop, semid, (int) nsops, 0, sops);
 }
